+++ src/Endpoint/AccountBrandingsEndpoint.php
@@ -12,7 +12,7 @@ use DigitalCz\DigiSign\Resource\ListResource;
 
 /**
  * @extends ResourceEndpoint<Branding>
- * @method ListResource<Branding> list(array $query = [])
+ * @method ListResource<Branding> list(array $query = [], bool $actions = true, bool $links = true)
  * @method Branding get(string $id)
  * @method Branding create(array $body)
  * @method Branding update(string $id, array $body)